move path_rewrite plugins into main package
