+++ plugin/path_rewrite/join_katakana_oov/Cargo.toml
@@ -9,6 +9,4 @@ license = "Apache-2.0"
 crate-type = ["cdylib"]
 
 [dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 sudachi = { path = "../../../sudachi" }
