+++ sudachi/src/plugin/path_rewrite/join_numeric/numeric_parser/mod.rs
@@ -25,7 +25,7 @@ pub enum Error {
     NONE,
     POINT,
     COMMA,
-    OTHER,
+    // OTHER,
 }
 
 /// Parses number written by arabic or kanji
